+++ src/DDTrace/Contracts/Tracer.php
@@ -127,4 +127,19 @@ interface Tracer
      * @return mixed
      */
     public function getPrioritySampling();
+
+    /**
+     * This behaves just like Tracer::startActiveSpan(), but it saves the Scope instance
+     * on the tracer to be accessed later by Tracer::getRootScope().
+     *
+     * @param string $operationName
+     * @param array $options
+     * @return Scope
+     */
+    public function startRootSpan($operationName, $options = []);
+
+    /**
+     * @return Scope|null
+     */
+    public function getRootScope();
 }