+++ tests/Common/TracerTestTrait.php
@@ -98,12 +98,20 @@ trait TracerTestTrait
         $curl =  curl_init(self::$agentRequestDumperUrl . '/replay');
         curl_setopt($curl, CURLOPT_RETURNTRANSFER, true);
         $response = curl_exec($curl);
+        if (!$response) {
+            return [];
+        }
 
         // For now we only support asserting traces against one dump at a time.
         $loaded = json_decode($response, true);
-        $rawTraces = json_decode($loaded['body'], true) ?: [];
 
+        if (!isset($loaded['body'])) {
+            return [];
+        }
+
+        $rawTraces = json_decode($loaded['body'], true);
         $traces = [];
+
         foreach ($rawTraces as $spansInTrace) {
             $spans = [];
             foreach ($spansInTrace as $rawSpan) {